[2707]-[General] Varias funciones y sprites
Se ha implementado el minimapa, el Game Over, el título, el HUD de unidad, el spawn de unidades, construccion de edificios con alguien ocupado de ellos, adición de recursos con el paso del tiempo una vez están completos y los enemigos atacando a las unidades aliadas.

Cosas a arreglar:
-Movimiento de los personajes sorteando obstáculos y parando en un lugar separados entre sí
-Zonas de respawn que no estén sobre árboles
-Gráficos unos sobre otros
-Cursor de construcción rojo o verde según si se puede construir
-IA de aliados cuando hay enemigos cerca

+++ Assets/Scripts/Personajes/Unidad/UnidadScriptable.cs
@@ -7,7 +7,7 @@ public class UnidadScriptable : ScriptableObject
 {
     public int vida;
     public int fuerza;
-    public int rangoAtaque;
+    public float rangoAtaque;
     public TipoUnidad tipo;
     public Bando bando;
     public float tiempoEnfriamientoAtaque;
